ADD - ID Identifiers for all Entitys. Managed to annotate with @MongoID to get the ids right in DB aswell.
Added functionality for deleting rides, rides will be deleted both in rideentity and in the user collection.

Added the abillity to add new bikes to user collection, and they will also appear in the bikeentity.

IF the user tries to add a bike with same Model, maker, year and colors as a already added bike, the existing bike will be added to the collection instead.